feat: modularize community page components - Extract LocationClimate into separate component - Update imports and layout - Improve code organization
