+++ app/(site)/community.tsx
@@ -1,13 +1,17 @@
 import { useTint } from '@tamagui/logo'
-import { Check, ChevronLeft, ChevronRight, Globe, Heart, Leaf, Palette, Sun, Users, X } from '@tamagui/lucide-icons'
-import { useMemo, useState } from 'react'
-import { Button, Card, Dialog, H1, H2, H3, Paragraph, Separator, Spacer, XStack, YStack } from 'tamagui'
+import { ChevronRight, Globe, Heart, Leaf, Palette, Sun, Users } from '@tamagui/lucide-icons'
+import { useMemo } from 'react'
+import { Button, Card, H1, H2, H3, Paragraph, Spacer, XStack, YStack } from 'tamagui'
 import { ContainerLarge } from '~/components/Containers'
 import { HeadInfo } from '~/components/HeadInfo'
 import { SocialLinksRow } from '~/features/site/home/SocialLinksRow'
 import { HomeSection, TintSection } from '~/features/site/home/TintSection'
 import { ThemeNameEffect } from '~/features/site/theme/ThemeNameEffect'
 import { FAQSection } from './components/FAQ'
+import { GallerySection } from './components/Gallery'
+import { LocationClimate } from './components/LocationClimate'
+import { MembershipTiers } from './components/MembershipTiers'
+import { NewsUpdates } from './components/NewsUpdates'
 import { teamMembers } from './data/team'
 
 export default function Community() {
@@ -212,172 +216,7 @@ export default function Community() {
       {/* Location & Climate */}
       <TintSection index={2}>
         <ContainerLarge position="relative">
-          <YStack zi={1} space="$6" mb="$8">
-            <YStack space="$6" mb="$6">
-              <H2 size="$9" ta="center" $sm={{ size: "$8" }}>Location & Climate</H2>
-              <Paragraph size="$6" ta="center" theme="alt2" maw={700} als="center" $sm={{ size: "$5" }}>
-                Experience the perfect blend of natural beauty and modern comfort in Mexico's stunning Riviera Maya
-              </Paragraph>
-            </YStack>
-
-            <XStack
-              pos="relative"
-              jc="center"
-              gap="$4"
-              $sm={{ fd: 'column' }}
-              width="100%"
-            >
-              <Card 
-                f={1} 
-                bw={1} 
-                bc="$borderColor" 
-                br="$6" 
-                elevation="$6" 
-                shadowRadius={60} 
-                width="48%" 
-                animation="medium"
-                pressStyle={{
-                  scale: 0.98,
-                  bc: "$color1",
-                }}
-                hoverStyle={{
-                  elevation: "$8",
-                  borderColor: "$color8",
-                  scale: 1.01
-                }}
-                $sm={{ width: '100%' }}
-              >
-                <YStack jc="center" p="$6" space="$6">
-                  <YStack space="$4">
-                    <H2 size="$8">Paradise Found</H2>
-                    <Paragraph size="$5" theme="alt2">
-                      Our carefully selected location offers the perfect blend of natural beauty and strategic advantages:
-                    </Paragraph>
-                  </YStack>
-                  <YStack space="$6">
-                    <XStack ai="center" space="$3">
-                      <Check size={20} color="var(--green10)" />
-                      <YStack space="$1">
-                        <Paragraph size="$4" fontWeight="600">Perfect Growing Climate</Paragraph>
-                        <Paragraph size="$3" theme="alt2">Year-round conditions ideal for tropical and subtropical crops</Paragraph>
-                      </YStack>
-                    </XStack>
-                    
-                    <XStack ai="center" space="$3">
-                      <Check size={20} color="var(--green10)" />
-                      <YStack space="$1">
-                        <Paragraph size="$4" fontWeight="600">Natural Resources</Paragraph>
-                        <Paragraph size="$3" theme="alt2">Crystal-clear cenotes and rich biodiversity on the property</Paragraph>
-                      </YStack>
-                    </XStack>
-
-                    <XStack ai="center" space="$3">
-                      <Check size={20} color="var(--green10)" />
-                      <YStack space="$1">
-                        <Paragraph size="$4" fontWeight="600">Strategic Location</Paragraph>
-                        <Paragraph size="$3" theme="alt2">Easy access to airports, beaches, and modern amenities</Paragraph>
-                      </YStack>
-                    </XStack>
-
-                    <XStack ai="center" space="$3">
-                      <Check size={20} color="var(--green10)" />
-                      <YStack space="$1">
-                        <Paragraph size="$4" fontWeight="600">Community Integration</Paragraph>
-                        <Paragraph size="$3" theme="alt2">Thriving expat community and local cultural experiences</Paragraph>
-                      </YStack>
-                    </XStack>
-
-                    <XStack ai="center" space="$3">
-                      <Check size={20} color="var(--green10)" />
-                      <YStack space="$1">
-                        <Paragraph size="$4" fontWeight="600">Development Ready</Paragraph>
-                        <Paragraph size="$3" theme="alt2">All permits secured and infrastructure planning complete</Paragraph>
-                      </YStack>
-                    </XStack>
-                  </YStack>
-                </YStack>
-              </Card>
-
-              <Card 
-                f={1} 
-                bw={1} 
-                bc="$borderColor" 
-                br="$6" 
-                elevation="$6" 
-                shadowRadius={60} 
-                width="48%" 
-                animation="medium"
-                pressStyle={{
-                  scale: 0.98,
-                  bc: "$color1",
-                }}
-                hoverStyle={{
-                  elevation: "$8",
-                  borderColor: "$color8",
-                  scale: 1.01
-                }}
-                $sm={{ width: '100%' }}
-              >
-                <YStack jc="center" p="$6" space="$6">
-                  <YStack space="$4">
-                    <H2 size="$8">Location Details</H2>
-                    <YStack space="$5">
-                      <YStack space="$3">
-                        <H3 size="$5" theme="alt2">Climate & Environment</H3>
-                        <YStack space="$2">
-                          <Paragraph size="$4">
-                            • Average temperature: 24-29°C (75-84°F)
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • Rainy season: June to October
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • Humidity: Moderate tropical climate
-                          </Paragraph>
-                        </YStack>
-                      </YStack>
-
-                      <YStack space="$3">
-                        <H3 size="$5" theme="alt2">Travel & Access</H3>
-                        <YStack space="$2">
-                          <Paragraph size="$4">
-                            • 45 min from Cancún International Airport
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • 20 min to nearest major hospital
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • 15 min to local markets and amenities
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • 10 min to pristine beaches
-                          </Paragraph>
-                        </YStack>
-                      </YStack>
-
-                      <YStack space="$3">
-                        <H3 size="$5" theme="alt2">Property Overview</H3>
-                        <YStack space="$2">
-                          <Paragraph size="$4">
-                            • 50 acres of pristine tropical land
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • Multiple natural cenotes on-site
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • Mix of jungle and clearings
-                          </Paragraph>
-                          <Paragraph size="$4">
-                            • Natural elevation changes
-                          </Paragraph>
-                        </YStack>
-                      </YStack>
-                    </YStack>
-                  </YStack>
-                </YStack>
-              </Card>
-            </XStack>
-          </YStack>
+          <LocationClimate />
         </ContainerLarge>
       </TintSection>
 
@@ -404,160 +243,7 @@ export default function Community() {
               Secure your place in our community with blockchain-powered ownership and exclusive benefits
             </Paragraph>
           </YStack>
-          <XStack flexWrap="wrap" gap="$4" jc="center" $sm={{ gap: "$3" }}>
-            <YStack
-              f={1}
-              miw={300}
-              maxWidth={350}
-              p="$6"
-              br="$4"
-              space="$4"
-              borderWidth={2}
-              borderColor="$borderColor"
-              pressStyle={{
-                scale: 0.98,
-                bc: "$color1",
-              }}
-              animation="medium"
-              hoverStyle={{
-                elevation: "$8",
-                borderColor: "$color8"
-              }}
-              $sm={{ 
-                miw: "auto",
-                w: "100%",
-                p: "$4"
-              }}
-            >
-              <YStack space="$2">
-                <H3 size="$6">Founding Member</H3>
-                <Paragraph theme="alt2" size="$8" fontWeight="800">5 ETH</Paragraph>
-              </YStack>
-              <Separator />
-              <YStack space="$3">
-                {[
-                  'Private villa allocation',
-                  'Full voting rights',
-                  'Revenue sharing',
-                  'Priority access to facilities',
-                  'Founding member NFT'
-                ].map(benefit => (
-                  <XStack key={benefit} ai="center" space="$3">
-                    <Check size={16} color="var(--color)" />
-                    <Paragraph size="$3">{benefit}</Paragraph>
-                  </XStack>
-                ))}
-              </YStack>
-              <Button 
-                size="$4" 
-                theme="green_alt2"
-                pressStyle={{
-                  scale: 0.97,
-                }}
-                animation="quick"
-                hoverStyle={{
-                  opacity: 0.9,
-                  scale: 1.02
-                }}
-              >
-                Join Waitlist
-              </Button>
-            </YStack>
-
-            <YStack
-              f={1}
-              miw={300}
-              maxWidth={350}
-              p="$6"
-              br="$4"
-              space="$4"
-              borderWidth={2}
-              borderColor="$borderColor"
-              theme="purple"
-              pressStyle={{
-                scale: 0.98,
-                bc: "$color1",
-              }}
-              animation="medium"
-              hoverStyle={{
-                elevation: "$8",
-                borderColor: "$color8"
-              }}
-              $sm={{ 
-                miw: "auto",
-                w: "100%",
-                p: "$4"
-              }}
-            >
-              <YStack space="$2">
-                <H3 size="$6">Core Member</H3>
-                <Paragraph theme="alt2" size="$8" fontWeight="800">2 ETH</Paragraph>
-              </YStack>
-              <Separator />
-              <YStack space="$3">
-                {[
-                  'Shared living space',
-                  'Partial voting rights',
-                  'Community participation',
-                  'Access to all facilities',
-                  'Core member NFT'
-                ].map(benefit => (
-                  <XStack key={benefit} ai="center" space="$3">
-                    <Check size={16} color="var(--color)" />
-                    <Paragraph size="$3">{benefit}</Paragraph>
-                  </XStack>
-                ))}
-              </YStack>
-              <Button size="$4" theme="purple_alt2">Reserve Spot</Button>
-            </YStack>
-
-            <YStack
-              f={1}
-              miw={300}
-              maxWidth={350}
-              p="$6"
-              br="$4"
-              space="$4"
-              borderWidth={2}
-              borderColor="$borderColor"
-              theme="blue"
-              pressStyle={{
-                scale: 0.98,
-                bc: "$color1",
-              }}
-              animation="medium"
-              hoverStyle={{
-                elevation: "$8",
-                borderColor: "$color8"
-              }}
-              $sm={{ 
-                miw: "auto",
-                w: "100%",
-                p: "$4"
-              }}
-            >
-              <YStack space="$2">
-                <H3 size="$6">Community Member</H3>
-                <Paragraph theme="alt2" size="$8" fontWeight="800">0.5 ETH</Paragraph>
-              </YStack>
-              <Separator />
-              <YStack space="$3">
-                {[
-                  'Access to common areas',
-                  'Community events',
-                  'Seasonal stays',
-                  'Basic amenities',
-                  'Community NFT'
-                ].map(benefit => (
-                  <XStack key={benefit} ai="center" space="$3">
-                    <Check size={16} color="var(--color)" />
-                    <Paragraph size="$3">{benefit}</Paragraph>
-                  </XStack>
-                ))}
-              </YStack>
-              <Button size="$4" theme="blue_alt2">Join Community</Button>
-            </YStack>
-          </XStack>
+          <MembershipTiers />
         </ContainerLarge>
       </TintSection>
 
@@ -702,104 +388,7 @@ export default function Community() {
               </Paragraph>
             </YStack>
 
-            <XStack flexWrap="wrap" gap="$4" jc="center">
-              {[
-                {
-                  date: 'January 2024',
-                  title: 'Land Survey Complete',
-                  description: 'Detailed topographical analysis reveals perfect building sites and confirms three pristine cenotes on the property.',
-                  tag: 'Development'
-                },
-                {
-                  date: 'December 2023',
-                  title: 'Community Design Workshop',
-                  description: 'Successful virtual workshop with future residents to finalize sustainable housing designs and community spaces.',
-                  tag: 'Community'
-                },
-                {
-                  date: 'November 2023',
-                  title: 'Solar Assessment',
-                  description: 'Technical study confirms excellent solar potential, enabling 100% renewable energy coverage for the community.',
-                  tag: 'Infrastructure'
-                },
-                {
-                  date: 'October 2023',
-                  title: 'Local Partnership',
-                  description: 'Agreement signed with local organic farmers to support our permaculture development and knowledge exchange.',
-                  tag: 'Partnership'
-                }
-              ].map((update, i) => (
-                <Card
-                  key={i}
-                  bw={1}
-                  bc="$borderColor"
-                  br="$6"
-                  elevation="$4"
-                  width="48%"
-                  p="$5"
-                  space="$3"
-                  pressStyle={{
-                    scale: 0.98,
-                    bc: "$color1",
-                  }}
-                  animation="medium"
-                  hoverStyle={{
-                    elevation: "$8",
-                    borderColor: "$color8",
-                    scale: 1.01
-                  }}
-                  $gtSm={{ minWidth: 400 }}
-                  $sm={{ width: '100%' }}
-                >
-                  <XStack jc="space-between" ai="center">
-                    <Paragraph size="$3" theme="alt2">{update.date}</Paragraph>
-                    <YStack 
-                      backgroundColor={
-                        update.tag === 'Development' ? '$green5' :
-                        update.tag === 'Community' ? '$blue5' :
-                        update.tag === 'Infrastructure' ? '$orange5' :
-                        '$purple5'
-                      }
-                      px="$2"
-                      py="$1"
-                      br="$4"
-                    >
-                      <Paragraph 
-                        size="$2" 
-                        color={
-                          update.tag === 'Development' ? '$green10' :
-                          update.tag === 'Community' ? '$blue10' :
-                          update.tag === 'Infrastructure' ? '$orange10' :
-                          '$purple10'
-                        }
-                      >
-                        {update.tag}
-                      </Paragraph>
-                    </YStack>
-                  </XStack>
-                  <H3 size="$6">{update.title}</H3>
-                  <Paragraph size="$4" theme="alt2">{update.description}</Paragraph>
-                </Card>
-              ))}
-            </XStack>
-
-            <YStack ai="center" mt="$4">
-              <Button 
-                size="$4" 
-                theme="alt2"
-                iconAfter={ChevronRight}
-                pressStyle={{
-                  scale: 0.97,
-                }}
-                animation="quick"
-                hoverStyle={{
-                  opacity: 0.9,
-                  scale: 1.02
-                }}
-              >
-                View All Updates
-              </Button>
-            </YStack>
+            <NewsUpdates />
           </YStack>
         </ContainerLarge>
       </TintSection>
@@ -948,271 +537,4 @@ export default function Community() {
 function CommunityLayout({ children }: { children: any }) {
   const { tint } = useTint()
   return <YStack theme={tint as any}>{useMemo(() => children, [children])}</YStack>
-}
-
-const galleryData = {
-  "Natural Cenotes": {
-    images: [
-      { url: "/images/gallery/cenote.jpg", caption: "Crystal-clear waters perfect for swimming" },
-      { url: "/images/gallery/cenote-2.jpg", caption: "Ancient limestone formations create natural pools" },
-      { url: "/images/gallery/cenote-3.jpg", caption: "Underground water systems feed our cenotes" }
-    ]
-  },
-  "Lush Jungle": {
-    images: [
-      { url: "/images/gallery/jungle.jpg", caption: "50 acres of pristine tropical forest" },
-      { url: "/images/gallery/jungle-2.jpg", caption: "Rich biodiversity with native species" },
-      { url: "/images/gallery/jungle-3.jpg", caption: "Natural trails through the property" }
-    ]
-  },
-  "Community Spaces": {
-    images: [
-      { url: "/images/gallery/community.jpg", caption: "Shared areas for connection & creativity" },
-      { url: "/images/gallery/community-2.jpg", caption: "Open-air gathering spaces" },
-      { url: "/images/gallery/community-3.jpg", caption: "Collaborative work areas" }
-    ]
-  },
-  "Sustainable Villas": {
-    images: [
-      { url: "/images/gallery/villa.jpg", caption: "Eco-friendly homes with modern amenities" },
-      { url: "/images/gallery/villa-2.jpg", caption: "Solar-powered living spaces" },
-      { url: "/images/gallery/villa-3.jpg", caption: "Natural building materials" }
-    ]
-  },
-  "Nearby Beaches": {
-    images: [
-      { url: "/images/gallery/beach.jpg", caption: "Just minutes from pristine coastline" },
-      { url: "/images/gallery/beach-2.jpg", caption: "Crystal clear Caribbean waters" },
-      { url: "/images/gallery/beach-3.jpg", caption: "White sand beaches nearby" }
-    ]
-  }
-}
-
-function GallerySection() {
-  const [selectedCategory, setSelectedCategory] = useState<string | null>(null)
-
-  return (
-    <>
-      <XStack flexWrap="wrap" gap="$4" jc="center">
-        {Object.entries(galleryData).map(([category, data], index) => (
-          <Card
-            key={category}
-            bw={1}
-            bc="$borderColor"
-            br="$6"
-            elevation="$4"
-            width={index >= 3 ? "48%" : "31%"}
-            height={index >= 3 ? 400 : 300}
-            pressStyle={{
-              scale: 0.98,
-              bc: "$color1",
-            }}
-            animation="medium"
-            hoverStyle={{
-              elevation: "$8",
-              borderColor: "$color8",
-              scale: 1.02
-            }}
-            onPress={() => setSelectedCategory(category)}
-            $gtSm={{ minWidth: index >= 3 ? 450 : 300 }}
-            $sm={{ width: '100%', height: index >= 3 ? 250 : 200 }}
-            style={{
-              backgroundImage: `url(${data.images[0].url})`,
-              backgroundSize: 'cover',
-              backgroundPosition: 'center',
-            }}
-          >
-            <YStack f={1} jc="flex-end" p="$4" space="$2" style={{
-              background: 'linear-gradient(0deg, rgba(0,0,0,0.7) 0%, rgba(0,0,0,0) 100%)',
-            }}>
-              <H3 size="$6" color="white">{category}</H3>
-              <Paragraph size="$3" color="white" o={0.9}>{data.images[0].caption}</Paragraph>
-            </YStack>
-          </Card>
-        ))}
-      </XStack>
-
-      {selectedCategory && (
-        <GalleryModal
-          isOpen={!!selectedCategory}
-          onClose={() => setSelectedCategory(null)}
-          category={selectedCategory}
-          images={galleryData[selectedCategory].images}
-        />
-      )}
-    </>
-  )
-}
-
-const GalleryModal = ({ 
-  isOpen, 
-  onClose, 
-  category,
-  images 
-}: { 
-  isOpen: boolean
-  onClose: () => void
-  category: string
-  images: { url: string; caption: string }[]
-}) => {
-  const [currentIndex, setCurrentIndex] = useState(0)
-  const [currentCategory, setCurrentCategory] = useState(category)
-
-  const handleNext = () => {
-    if (currentIndex < images.length - 1) {
-      setCurrentIndex(currentIndex + 1)
-    } else {
-      const categories = Object.keys(galleryData)
-      const currentCategoryIndex = categories.indexOf(currentCategory)
-      if (currentCategoryIndex < categories.length - 1) {
-        const nextCategory = categories[currentCategoryIndex + 1]
-        setCurrentCategory(nextCategory)
-        setCurrentIndex(0)
-      } else {
-        setCurrentCategory(categories[0])
-        setCurrentIndex(0)
-      }
-    }
-  }
-
-  const handlePrevious = () => {
-    if (currentIndex > 0) {
-      setCurrentIndex(currentIndex - 1)
-    } else {
-      const categories = Object.keys(galleryData)
-      const currentCategoryIndex = categories.indexOf(currentCategory)
-      if (currentCategoryIndex > 0) {
-        const prevCategory = categories[currentCategoryIndex - 1]
-        setCurrentCategory(prevCategory)
-        setCurrentIndex(galleryData[prevCategory].images.length - 1)
-      } else {
-        const lastCategory = categories[categories.length - 1]
-        setCurrentCategory(lastCategory)
-        setCurrentIndex(galleryData[lastCategory].images.length - 1)
-      }
-    }
-  }
-
-  return (
-    <Dialog modal open={isOpen} onOpenChange={(open) => !open && onClose()}>
-      <Dialog.Portal>
-        <Dialog.Overlay
-          key="overlay"
-          animation="quick"
-          opacity={0.5}
-          enterStyle={{ opacity: 0 }}
-          exitStyle={{ opacity: 0 }}
-        />
-        <Dialog.Content
-          bordered
-          elevate
-          key="content"
-          animation={[
-            'quick',
-            {
-              opacity: {
-                overshootClamping: true,
-              },
-            },
-          ]}
-          enterStyle={{ x: 0, y: -20, opacity: 0, scale: 0.9 }}
-          exitStyle={{ x: 0, y: 10, opacity: 0, scale: 0.95 }}
-          x={0}
-          scale={1}
-          opacity={1}
-          y={0}
-          width="90%"
-          maxWidth={950}
-          height="90vh"
-          maxHeight={750}
-        >
-          <YStack space="$4" height="100%" padding="$4">
-            <XStack jc="space-between" ai="center">
-              <H2 size="$8">{currentCategory}</H2>
-              <Button 
-                size="$3" 
-                circular 
-                icon={X} 
-                onPress={onClose}
-                theme="alt2"
-              />
-            </XStack>
-
-            <YStack f={1} position="relative">
-              <Card
-                elevation={10}
-                br="$4"
-                bw={1}
-                width="100%"
-                height="100%"
-                style={{
-                  backgroundImage: `url(${galleryData[currentCategory].images[currentIndex].url})`,
-                  backgroundSize: 'cover',
-                  backgroundPosition: 'center',
-                  transition: 'background-image 0.3s ease-in-out'
-                }}
-              >
-                <YStack
-                  width="100%"
-                  height="100%"
-                  style={{
-                    backgroundImage: `url(${galleryData[currentCategory].images[currentIndex].url})`,
-                    backgroundSize: 'cover',
-                    backgroundPosition: 'center',
-                    transition: 'all 0.3s ease-in-out'
-                  }}
-                />
-              </Card>
-              
-              <XStack 
-                position="absolute" 
-                bottom="$4" 
-                left="50%" 
-                x="-50%" 
-                space="$4"
-                elevation={10}
-              >
-                <Button 
-                  size="$4" 
-                  circular 
-                  icon={ChevronLeft}
-                  onPress={handlePrevious}
-                  theme="alt2"
-                  backgroundColor="$background"
-                />
-                <Button 
-                  size="$4" 
-                  circular 
-                  icon={ChevronRight}
-                  onPress={handleNext}
-                  theme="alt2"
-                  backgroundColor="$background"
-                />
-              </XStack>
-            </YStack>
-
-            <YStack space="$4">
-              <Paragraph size="$4" ta="center" theme="alt2">
-                {galleryData[currentCategory].images[currentIndex].caption}
-              </Paragraph>
-              
-              <XStack jc="center" space="$2">
-                {galleryData[currentCategory].images.map((_, index) => (
-                  <YStack
-                    key={index}
-                    width={8}
-                    height={8}
-                    borderRadius={100}
-                    backgroundColor={index === currentIndex ? "$color" : "$color4"}
-                    pressStyle={{ scale: 0.9 }}
-                    onPress={() => setCurrentIndex(index)}
-                  />
-                ))}
-              </XStack>
-            </YStack>
-          </YStack>
-        </Dialog.Content>
-      </Dialog.Portal>
-    </Dialog>
-  )
 }
